requirements.txt
alembic==1.13.1
astroid==2.13.5
attrs==24.2.0
Authlib==1.3.2
bcrypt==4.2.0
black==23.1.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Deprecated==1.2.15
dill==0.3.9
dnspython==2.6.1
email_validator==2.2.0
flake8==5.0.4
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
# Flask-Dance==7.1.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-OAuthlib==0.9.5
flask-redis==0.4.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==20.1.0
idna==3.8
importlib_resources==6.4.5
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.3
lazy-object-proxy==1.10.0
limits==3.13.0
mailjet-rest==1.3.4
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
msgspec==0.18.6
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
oauthlib==2.1.0
ordered-set==4.1.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.9.1
pycparser==2.22
pyflakes==2.5.0
Pygments==2.18.0
PyJWT==2.9.0
pylint==2.15.3
PyMySQL==1.1.0
pytest==7.2.1
python-dotenv==1.0.1
redis==5.2.0
requests==2.28.1
requests-oauthlib==1.1.0
rich==13.9.4
setuptools==75.4.0
SQLAlchemy==2.0.29
tomlkit==0.13.2
typing_extensions==4.11.0
urllib3==1.26.20
URLObject==2.4.3
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.16.0
WTForms==3.1.2
